Module ferrous::dsl

The DSL module for the ferrous crate.

Usually, a user of this library would want to import all symbols from this module.

use ferrous::dsl::*;

expect(&1).to(equal(&1));

Traits

Assertion
Trait for assertion types.

AsyncAssertion
Trait mapping to/not_to to should/should_not for better readability.

Matcher
Trait that must be implemented by any matcher type.

Functions

be_err
Matcher that asserts a value is a Err variant.

be_none
Matcher that asserts a value is a None variant.

be_ok
Matcher that asserts a value is a Ok variant.

be_some
Matcher that asserts a value is a Some variant.

consistently
Asynchronous assertion that asserts the given closure consistently returns a value that satisfies the given matcher.

consistently_with_timeout
Asynchronous consistent assertion with user-defined timeout.

contain
Container matcher that asserts a given element exists within an iterator.

equal
Simple equality matcher

eventually
Asynchronous assertion that asserts the given closure eventually returns a value that satisfies the given matcher.

eventually_with_timeout
Asynchronous eventual assertion with user-defined timeout.

expect
Simple expectation assertion.