Trait ferrous::dsl::Matcher

pub trait Matcher<A> {
    fn matches(&self, actual: &A) -> bool;
    fn failure_message(&self, actual: &A) -> String;
    fn negated_failure_message(&self, actual: &A) -> String;
}

Trait that must be implemented by any matcher type.

Required Methods

fn matches(&self, actual: &A) -> bool

Defines whether or not the actual value matches the expected value

fn failure_message(&self, actual: &A) -> String

The positive failure message

fn negated_failure_message(&self, actual: &A) -> String

The megative failure message